treestatus/api: rename from releng_treestatus
