+++ nix/setup.cfg
@@ -19,6 +19,9 @@ sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 known_first_party =
     backend_common
     cli_common
+    code_coverage_backend
+    code_coverage_bot
+    code_coverage_crawler
     releng_archiver
     releng_clobberer
     releng_mapper
@@ -26,18 +29,15 @@ known_first_party =
     releng_notification_policy
     releng_tokens
     releng_tooltool
-    releng_treestatus
-    code_coverage_bot
-    code_coverage_backend
-    code_coverage_crawler
+    shipit_api
     shipit_pipeline
     pulselistener
     shipit_signoff
     shipit_taskcluster
-    shipit_api
     static_analysis_bot
-    uplift_bot
+    treestatus_api
     uplift_backend
+    uplift_bot
 
 [mypy]
 # Specify the target platform details in config, so your developers are
