+++ src/docs/deploy/heroku-target.rst
@@ -31,12 +31,12 @@ To do a manual deployment you need to:
 - and run ``./please tools deploy:HEROKU ...``
 
 You can check all possible options of this command by appending ``--help`` at
-the end. An example command to deploy ``releng-treestatus`` project to
+the end. An example command to deploy ``treestatus/api`` project to
 production environment would be:
 
 .. code-block:: console
 
-    $ ./please tools deploy:HEROKU releng-treestatus \
+    $ ./please tools deploy:HEROKU treestatus/api \
             --heroku-app="releng-production-treestatus" \
             --extra-attribute=".deploy.production" \
             --taskcluster-secret="repo:github.com/mozilla-releng/services:branch:production" \
@@ -62,32 +62,32 @@ Testing docker image locally
 Above mentioned command would internally build docker image. It is also
 possible to build docker image and run it locally.
 
-To build docker image for ``releng-treestatus`` project do:
+To build docker image for ``treestatus/api`` project do:
 
 .. code-block:: console
 
-    $ ./please tools build releng-treestatus \
+    $ ./please tools build treestatus/api \
         --extra-attribute="docker"
-    $ realpath tmp/result-build-releng-treestatus-1
-    /nix/store/...-docker-image-mozilla-releng-treestatus.tar.gz
+    $ realpath tmp/result-build-treestatus-api-1
+    /nix/store/...-docker-image-mozilla-treestatus-api.tar.gz
 
-A symlink to Docker image is create at ``tmp/result-build-releng-treestatus-1``
+A symlink to Docker image is create at ``tmp/result-build-treestatus-api-1``
 location.
 
 Now all you need to do is load image and run it:
 
 .. code-block:: console
 
-    $ cat tmp/result-build-releng-treestatus-1 | docker load
+    $ cat tmp/result-build-treestatus-api-1 | docker load
     $ docker images
     ...
-    mozilla-releng-treestatus   1.0.0   7a7c7882c4a3   47 years ago   262 MB
+    mozilla-treestatus-api   1.0.0   7a7c7882c4a3   47 years ago   262 MB
     ...
     $ docker run \
         -e TASKCLUSTER_SECRET=repo:github.com/mozilla-releng/services:branch:staging \
         -e TASKCLUSTER_CLIENT_ID=... \
         -e TASKCLUSTER_ACCESS_TOKEN=... \
-            mozilla-releng-treestatus:1.0.0
+            mozilla-treestatus-api:1.0.0
 
 
 .. _`Takcluster's Client Manager`: https://tools.taskcluster.net/auth/clients/
