+++ src/releng_frontend/src/index.js
@@ -31,7 +31,7 @@ var init = function() {
     // Start the ELM application
     var app = require('./Main.elm').Main.fullscreen({
       auth0: localstorage.load_item(AUTH_KEY),
-      treestatusUrl: getData('releng-treestatus-url', process.env.RELENG_TREESTATUS_URL),
+      treestatusUrl: getData('treestatus-api-url', process.env.RELENG_TREESTATUS_URL),
       docsUrl: getData('docs-url', process.env.DOCS_URL),
       version: release_version,
       identityUrl: getData('releng-notification-identity-url', process.env.RELENG_NOTIFICATION_IDENTITY_URL),
