+++ src/releng_frontend/src/static/tokens/index.html
@@ -187,7 +187,7 @@
   <script src="../tcauth.js" type="text/javascript"></script>
   <script src="tokens.js" type="text/javascript"></script>
   <script type="text/javascript">
-    window.tcauth.setup('tokens', 'https://localhost:8003');
+    window.tcauth.setup('releng-tokens', 'https://localhost:8003');
   </script>
 </body>
 </html>
