+++ src/releng_frontend/src/static/tooltool/index.html
@@ -124,7 +124,7 @@
   <script src="../tcauth.js" type="text/javascript"></script>
   <script src="tooltool.js" type="text/javascript"></script>
   <script type="text/javascript">
-    window.tcauth.setup('tooltool', 'https://localhost:8002');
+    window.tcauth.setup('releng-tooltool', 'https://localhost:8002');
   </script>
 </body>
 </html>
