--- src/releng_treestatus/requirements-dev.txt
@@ -1,1 +1,0 @@
--r ./../../lib/backend_common/requirements-dev.txt
