--- src/releng_treestatus/requirements_override.nix
@@ -1,1 +1,0 @@
-../../nix/requirements_override.nix
