--- src/releng_treestatus/setup.cfg
@@ -1,1 +1,0 @@
-../../nix/setup.cfg
