+++ src/treestatus/api/requirements-dev.txt
@@ -1,0 +1,1 @@
+-r ./../../../lib/backend_common/requirements-dev.txt
