+++ src/treestatus/api/requirements.nix
@@ -1277,7 +1277,7 @@ let
 
     "mozilla-backend-common" = python.mkDerivation {
       name = "mozilla-backend-common-1.0.0";
-      src = pkgs.lib.cleanSource ./../../lib/backend_common;
+      src = pkgs.lib.cleanSource ./../../../lib/backend_common;
       doCheck = commonDoCheck;
       checkPhase = "";
       installCheckPhase = "";
@@ -1316,7 +1316,7 @@ let
 
     "mozilla-cli-common" = python.mkDerivation {
       name = "mozilla-cli-common-1.0.0";
-      src = pkgs.lib.cleanSource ./../../lib/cli_common;
+      src = pkgs.lib.cleanSource ./../../../lib/cli_common;
       doCheck = commonDoCheck;
       checkPhase = "";
       installCheckPhase = "";
