+++ src/treestatus/api/requirements.txt
@@ -1,0 +1,7 @@
+-e ./../../../lib/cli_common[log] #egg=mozilla-cli-common
+-e ./../../../lib/backend_common[log,security,cors,api,auth,db,cache,pulse] #egg=mozilla-backend-common
+
+gunicorn
+psycopg2
+pytz
+redis
