+++ src/treestatus/api/treestatus_api/config.py
@@ -3,5 +3,5 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-PROJECT_NAME = 'releng-treestatus'
-APP_NAME = 'releng_treestatus'
+PROJECT_NAME = 'treestatus/api'
+APP_NAME = 'treestatus_api'
